Improved the LocalImageProcessingListeners error handling

+++ Event/LocalImageProcessingListener.php
@@ -3,55 +3,31 @@ App::uses('CakeEventListener', 'Event');
 /**
  * Local Image Processor Event Listener for the CakePHP FileStorage plugin
  *
+ * @author Florian Krämer
  * @copy 2012 Florian Krämer
  * @license MIT
  */
 class LocalImageProcessingListener implements CakeEventListener {
-
 /**
- * It is required to get the file first and write it to a tmp file
+ * Implemented Events
  *
- * The adapter might not be one that is using a local file system, so we first
- * get the file from the storage system, store it locally in a tmp file and 
- * later load the new file that was generated based on the tmp file into the 
- * storage adapter. This method here just generates the tmp file.
- *
- * @param 
- * @param 
- * @return 
+ * @return array
  */
-	protected function _tmpFile($Storage, $path) {
-		try {
-			$tmpFile = TMP . String::uuid();
-			$imageData = $Storage->read($path);
-			file_put_contents($tmpFile, $imageData);
-			return $tmpFile;
-		} catch (Exception $e) {
-			return false;
-		}
-	}
-
-/**
- * Check if the event can be processed
- *
- * @param CakeEvent $Event
- * @return boolean
- */
-	protected function _checkEvent($Event) {
-		$Model = $Event->subject();
-		return ($Model instanceOf ImageStorage && isset($Event->data['record'][$Model->alias]['adapter']) && $Event->data['record'][$Model->alias]['adapter'] == 'Local');
-	}
-
 	public function implementedEvents() {
 		return array(
 			'ImageStorage.createVersion' => 'createVersions',
 			'ImageStorage.removeVersion' => 'removeVersions',
 			'ImageStorage.afterSave' => 'afterSave',
 			'ImageStorage.afterDelete' => 'afterDelete'
-			//'ImageStorage.beforeDelete' => 'removeVersion'
 		);
 	}
 
+/**
+ * Creates versions for a given image record
+ *
+ * @param CakeEvent $Event
+ * @return void
+ */
 	public function createVersions($Event) {
 		if ($this->_checkEvent($Event)) {
 			$Model = $Event->subject();
@@ -60,17 +36,23 @@ class LocalImageProcessingListener implements CakeEventListener {
 
 			$tmpFile = $this->_tmpFile($Storage, $record['path']);
 
-			foreach ($Event->data['operations'] as $version => $operations) {
-				$hash = $Model->hashOperations($operations);
-				$string = substr($record['path'], 0, - (strlen($record['extension'])) -1);
-				$string .= '.' . $hash . '.' . $record['extension'];
+			try {
+				foreach ($Event->data['operations'] as $version => $operations) {
+					$hash = $Model->hashOperations($operations);
+					$string = substr($record['path'], 0, - (strlen($record['extension'])) -1);
+					$string .= '.' . $hash . '.' . $record['extension'];
 
-				if ($Storage->has($string)) {
-					return true;
+					if ($Storage->has($string)) {
+						continue;
+					}
+
+					$image = $Model->processImage($tmpFile, null, array('format' => $record['extension']), $operations);
+					$result = $Storage->write($string, $image->get($record['extension']), true);
 				}
-
-				$image = $Model->processImage($tmpFile, null, array('format' => $record['extension']), $operations);
-				$result = $Storage->write($string, $image->get($record['extension']), true);
+			} catch (Exception $e) {
+				$this->log($e->getMessage(), 'file_storage');
+				unlink($tmpFile);
+				throw $e;
 			}
 
 			unlink($tmpFile);
@@ -79,6 +61,12 @@ class LocalImageProcessingListener implements CakeEventListener {
 		}
 	}
 
+/**
+ * Removes versions for a given image record
+ *
+ * @param CakeEvent $Event
+ * @return void
+ */
 	public function removeVersions($Event) {
 		if ($this->_checkEvent($Event)) {
 			$Model = $Event->subject();
@@ -91,9 +79,11 @@ class LocalImageProcessingListener implements CakeEventListener {
 				$string .= '.' . $hash . '.' . $record['extension'];
 
 				try {
-					$Storage->delete($string);
+					if ($Storage->has($string)) {
+						$Storage->delete($string);
+					}
 				} catch (Exception $e) {
-					// No need to do anything here
+					$this->log($e->getMessage(), 'file_storage');
 				}
 			}
 
@@ -118,4 +108,41 @@ class LocalImageProcessingListener implements CakeEventListener {
 		}
 	}
 
+/**
+ * It is required to get the file first and write it to a tmp file
+ *
+ * The adapter might not be one that is using a local file system, so we first
+ * get the file from the storage system, store it locally in a tmp file and
+ * later load the new file that was generated based on the tmp file into the
+ * storage adapter. This method here just generates the tmp file.
+ *
+ * @param
+ * @param
+ * @return
+ */
+	protected function _tmpFile($Storage, $path) {
+		try {
+			if (!is_dir(TMP . 'image-processing')) {
+				mkdir(TMP . 'image-processing');
+			}
+
+			$tmpFile = TMP . 'image-processing' . DS . String::uuid();
+			$imageData = $Storage->read($path);
+			file_put_contents($tmpFile, $imageData);
+			return $tmpFile;
+		} catch (Exception $e) {
+			return false;
+		}
+	}
+
+/**
+ * Check if the event can be processed
+ *
+ * @param CakeEvent $Event
+ * @return boolean
+ */
+	protected function _checkEvent($Event) {
+		$Model = $Event->subject();
+		return ($Model instanceOf ImageStorage && isset($Event->data['record'][$Model->alias]['adapter']) && $Event->data['record'][$Model->alias]['adapter'] == 'Local');
+	}
 }
